Added use of font.scale property in JSHManager.properties.

+++ src/jshm/util/Properties.java
@@ -85,7 +85,7 @@ public class Properties extends java.util.Properties {
 		return defaultValue;
 	}
 	
-	public double getFloat(String propertyName) {
+	public float getFloat(String propertyName) {
 		float defaultValue = 0.0f;
 		
 		try {
